+++ src/hayate/__init__.py
@@ -14,7 +14,7 @@ from .urlpattern import URLPattern, URLPatternResult
 from .validator import validator
 from .websocket import WebSocket, WebSocketClosed
 
-__version__ = "0.3.2"
+__version__ = "0.4.0"
 
 __all__ = [
     "URL",